only copies the slides text files if they exist; this is to keep the script compatible with old recordings

+++ record-and-playback/presentation/scripts/process/presentation.rb
@@ -79,7 +79,9 @@ if not FileTest.directory?(target_dir)
            #BigBlueButton::Presentation.convert_pdf_to_png(pdf_page, "#{target_pres_dir}/slide-#{page}.png")
            command = "convert -density 300x300 -resize 1600x1200 -quality 90 +dither -depth 8 -colors 256 #{pdf_page} #{target_pres_dir}/slide-#{page}.png"
            BigBlueButton.execute(command)
-           FileUtils.cp("#{pres_dir}/textfiles/slide-#{page}.txt", "#{target_pres_dir}/textfiles")
+           if File.exist?("#{pres_dir}/textfiles/slide-#{page}.txt") then
+             FileUtils.cp("#{pres_dir}/textfiles/slide-#{page}.txt", "#{target_pres_dir}/textfiles")
+           end
          end
     else
         ext = File.extname("#{images[0]}")
